Use .NET8 and enable AOT compatibility for the main project. Move Utils into extensions project.
