+++ Exemples/SDL_EXTENSIONS/SDL_EXTENSIONS.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Platforms>x64;AnyCPU;x86</Platforms>
     <ApplicationIcon />
     <StartupObject />
@@ -32,6 +32,10 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\SDL-Sharp\SDL-Sharp.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="protobuf-net" Version="3.2.30" />
   </ItemGroup>
 
   <ItemGroup>
